// clients/typescript/src/API.ts
export type CreateKudoInput = {
  id?: string | null,
  giverId: string,
  receiverId: string,
  message: string,
  kudoVerb: KudoVerb,
  dataSourceApp: DataSourceApp,
  createdAt?: string | null,
  type?: string | null,
  link: string,
  metadata?: string | null,
};

export enum KudoVerb {
  kudos = "kudos",
  props = "props",
  sparkles = "sparkles",
  fireworks = "fireworks",
}


export enum DataSourceApp {
  github = "github",
  slack = "slack",
  teams = "teams",
  twitter = "twitter",
  other = "other",
}


export type ModelKudoConditionInput = {
  giverId?: ModelIDInput | null,
  receiverId?: ModelIDInput | null,
  message?: ModelStringInput | null,
  kudoVerb?: ModelKudoVerbInput | null,
  dataSourceApp?: ModelDataSourceAppInput | null,
  createdAt?: ModelStringInput | null,
  type?: ModelStringInput | null,
  link?: ModelStringInput | null,
  metadata?: ModelStringInput | null,
  and?: Array< ModelKudoConditionInput | null > | null,
  or?: Array< ModelKudoConditionInput | null > | null,
  not?: ModelKudoConditionInput | null,
};

export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type ModelKudoVerbInput = {
  eq?: KudoVerb | null,
  ne?: KudoVerb | null,
};

export type ModelDataSourceAppInput = {
  eq?: DataSourceApp | null,
  ne?: DataSourceApp | null,
};

export type Kudo = {
  __typename: "Kudo",
  id: string,
  giverId: string,
  receiverId: string,
  message: string,
  kudoVerb: KudoVerb,
  dataSourceApp: DataSourceApp,
  createdAt: string,
  type?: string | null,
  link: string,
  metadata?: string | null,
  giver?: Person | null,
  receiver?: Person | null,
  updatedAt: string,
};

export type Person = {
  __typename: "Person",
  id: string,
  username: string,
  email?: string | null,
  profileUrl?: string | null,
  profileImageUrl?: string | null,
  dataSourceApp: DataSourceApp,
  kudosGiven?: ModelKudoConnection | null,
  kudosReceived?: ModelKudoConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelKudoConnection = {
  __typename: "ModelKudoConnection",
  items:  Array<Kudo | null >,
  nextToken?: string | null,
};

export type UpdateKudoInput = {
  id: string,
  giverId?: string | null,
  receiverId?: string | null,
  message?: string | null,
  kudoVerb?: KudoVerb | null,
  dataSourceApp?: DataSourceApp | null,
  createdAt?: string | null,
  type?: string | null,
  link?: string | null,
  metadata?: string | null,
};

export type DeleteKudoInput = {
  id: string,
};

export type CreatePersonInput = {
  id?: string | null,
  username: string,
  email?: string | null,
  profileUrl?: string | null,
  profileImageUrl?: string | null,
  dataSourceApp: DataSourceApp,
};

export type ModelPersonConditionInput = {
  username?: ModelStringInput | null,
  email?: ModelStringInput | null,
  profileUrl?: ModelStringInput | null,
  profileImageUrl?: ModelStringInput | null,
  dataSourceApp?: ModelDataSourceAppInput | null,
  and?: Array< ModelPersonConditionInput | null > | null,
  or?: Array< ModelPersonConditionInput | null > | null,
  not?: ModelPersonConditionInput | null,
};
// ...
export type CreateKudoMutationVariables = {
  input: CreateKudoInput,
  condition?: ModelKudoConditionInput | null,
};
// ...
export type CreatePersonMutationVariables = {
  input: CreatePersonInput,
  condition?: ModelPersonConditionInput | null,
};